demo/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
configparser==5.0.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-allauth==0.41.0
django-bootstrap==0.2.4
django-bootstrap3==14.0.0
django-filter==2.3.0

django-markdown-deux==1.0.5
django-ranged-response==0.2.0
django-registration-redux==2.8
django-simple-captcha==0.5.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
markdown2==2.3.9

oauthlib==3.1.0
Pillow==7.1.2

PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.8
zipp==3.1.0
psycopg2